+++ piklist/parts/settings/system-status-settings.php
@@ -78,13 +78,63 @@ function get_development_data(){
     return $short_data.'<br/><hr/>'.$current_theme.$parent_theme;
 }
 
+/**
+ * Get size information for all currently-registered image sizes.
+ *
+ * @global $_wp_additional_image_sizes
+ * @uses   get_intermediate_image_sizes()
+ * @return array $sizes Data for all currently-registered image sizes.
+ */
+function get_image_sizes() {
+    global $_wp_additional_image_sizes;
+
+    $sizes = array();
+    $output = '';
+
+    foreach ( get_intermediate_image_sizes() as $_size ) {
+        if ( in_array( $_size, array('thumbnail', 'medium', 'medium_large', 'large') ) ) {
+            $sizes[ $_size ]['width']  = get_option( "{$_size}_size_w" );
+            $sizes[ $_size ]['height'] = get_option( "{$_size}_size_h" );
+            $sizes[ $_size ]['crop']   = (bool) get_option( "{$_size}_crop" );
+        } elseif ( isset( $_wp_additional_image_sizes[ $_size ] ) ) {
+            $sizes[ $_size ] = array(
+                'width'  => $_wp_additional_image_sizes[ $_size ]['width'],
+                'height' => $_wp_additional_image_sizes[ $_size ]['height'],
+                'crop'   => $_wp_additional_image_sizes[ $_size ]['crop'],
+            );
+        }
+    }
+
+
+     foreach ($sizes as $image_size => $image_size_data){
+
+
+         $crop = '';  
+
+         if(is_array($image_size_data['crop'])){
+            $crop = 'true '.$image_size_data['crop'][0].'-'.$image_size_data['crop'][0];
+         }else{
+            $crop = $image_size_data['crop'] ? 'true auto' : 'false';
+         }
+
+     $output .= '<tr><td>'. $image_size .'</td><td> '. $image_size_data['width'] .'x'. $image_size_data['height'] .'</td><td>Croop: '. $crop .'</td></tr> '; 
+    }
+
+    return '<table>'. $output . '</table>';
+}
+
 // Check favicon existence.
-$ico_path = get_stylesheet_directory() .'/favicon.ico';	
+$ico_path = get_stylesheet_directory() .'/favicon.ico'; 
 $favicon = false !== get_transient('site_favicon') ? get_transient('site_favicon') : '<span class="wp-ui-text-notification dashicons dashicons-warning"></span> No data';
 
 $favicon_status = wps_file_exist( $ico_path ) ? '<span class="dashicons dashicons-yes wp-ui-text-highlight"></span>':'<span class="wp-ui-text-notification dashicons dashicons-warning"></span> Missing';
 
 echo get_development_data();
 
+echo '<br/>';
+
 echo 'Favicon in theme root: '.$favicon_status.' | transient: '. $favicon;
 
+echo '<br/>';
+
+echo 'Available image sizes: '.get_image_sizes();